Add new networking tables for enhancements to IP managment
Three new tables: net_default_maps, net_default_ips and net_default_trunks
will be used to allow for management of multiple interfaces as well as
multiple IPs on a given interface, along with a few other enhancements
to be used by various tools.

This is the initial commit and does not have all the relationships in
place yet.