Redo model in Declarative
To make it easier to normalize the naming scheme in the database schema,
the model for TagOpsDB has been reverted back to Declarative from elixir
and the version of SQLAlchemy bumped to 0.9.4.  The new constraint naming
isn't yet in place and none of the deploy code in the library or TDS has
been adjusted to the new model yet, but the new model is functional at
the moment.

Several changes were also made to match the current production database
and a few latent issues with the model were found and corrected (mainly
in several relationships).  The deploy modules were also reverted to use
the original Declarative setup.