Set up Alembic for schema/model changes
Initial use of Alembic to maintain schema changes in the database that
will be mirrored in the model for the library.

This contains the first change to be made (fully utilizing the new
environments table - just the addition of the column to allow scripts
to be corrected before removing the old 'environment' columns).