Remove 'environment' column from tables
With the new 'environment_id' column (and 'environment_obj' relationship)
in place, adding a new alembic script to remove the 'environment' column
from the app_deployments and hosts tables, as well as adding a (temporary)
property for 'environment' on each.

This change also contains one minor fix to the previous alembic script.