SRE-15125: switch to PyMySQL
The MySQLdb module is unmaintained. There is a drop-in replacement:
"mysqlclient"; unfortunately, this does not exist in CentOS for python
2.

Using PyMySQL, at least for now, should give us forward/backward
compatibility.

This requires setting the charset to 'utf8'--otherwise, the driver (or
the database?) tries to use 'utf8mb4' and fails.