SRE-21944: fix discovery of deployments needing validation
1. Rather than finding the latest deployments that are not validated,
find the latest deployments and return the ones that are not validated.
The difference is that the previous approach returns deployments which
were never validated but which have been superseded by newer
deployments. Fixing this stops the false positives for deployments which
have been superseded.

2. Stop using the "group by trick". This was a handy-dandy way to find
rows of a table for which a given column was the maximum (or minimum),
and was promoted in examples around at least the early 2010s. This was
non-standard SQL, though, and the value of non-aggregated columns was
undefined behavior. MariaDB has apparently changed that bevavior.
https://mariadb.com/kb/en/group-by-trick-has-been-optimized-away/
Fixing this reveals some deployments which need to be validated but
which were previously not found.

3. Stop using AppDefinition.status. We don't keep this column up to
date. Instead, see if there are any hosts. Any appType with hosts is
active. This avoids notifying for appTypes which no longer have any
hosts.