SRE-21944: remove unused methods
find_deployed_version() and find_app_deployment() have group_by() with
non-aggregated columns, which is undefined behavior. Thankfully, they
are unused, so we don't need to figure out if we were relying on the
undefined behavior.

find_app_deployment() exists in TDS but is in an unused method which I
am also removing.